[#18] refactor : Logger 파일명 변경, 패키지 이동
