+++ src/main/kotlin/com/flabedu/blackpostoffice/commom/utils/constants/LoggerContants.kt
@@ -1,4 +1,4 @@
-package com.flabedu.blackpostoffice.commom.logging
+package com.flabedu.blackpostoffice.commom.utils.constants
 
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
